+++ lib/helpers/index.d.ts
@@ -1,3 +1,5 @@
-import { MdList } from '../types';
+import * as FormData from 'form-data';
+import { DiscordMessage, MdList } from '../types';
 export declare function getDiscordColor(hexColor: string): number;
 export declare function convertToList(item: string | MdList): MdList;
+export declare function getFormData(data: DiscordMessage): FormData;