+++ lib/index.d.ts
@@ -2,5 +2,5 @@ import HookBuilder from './services/HookBuilder';
 import MessageBuilder from './services/MessageBuilder';
 import EmbedBuilder from './services/EmbedBuilder';
 import MarkdownBuilder from './services/MarkdownBuilder';
-import { DiscordMessage, Embed, Author, Field, Footer, Image, TTS, Parse, AllowedMentions } from './types';
-export { HookBuilder, MessageBuilder, EmbedBuilder, MarkdownBuilder, DiscordMessage, Embed, Author, Field, Footer, Image, TTS, AllowedMentions, Parse, };
+import { DiscordMessage, Embed, Author, Field, Footer, Image, TTS, Parse, AllowedMentions, WebhookSettings, DiscordFile } from './types';
+export { HookBuilder, MessageBuilder, EmbedBuilder, MarkdownBuilder, DiscordMessage, Embed, Author, Field, Footer, Image, TTS, AllowedMentions, WebhookSettings, DiscordFile, Parse, };