+++ lib/services/EmbedBuilder.d.ts
@@ -1,6 +1,7 @@
 import { Embed, Author, Field, Footer } from '../types';
 export default class EmbedBuilder {
     private embed;
+    private isAllCharactersValid;
     setTitle(title?: string): this;
     setDescription(description?: string): this;
     setUrl(url?: string): this;
@@ -8,8 +9,8 @@ export default class EmbedBuilder {
     addField(field: Field): this;
     setAuthor(author?: Author): this;
     setFooter(footer?: Footer): this;
-    setImage(url?: string): this;
-    setThumbnail(url?: string): this;
+    setImage(url?: string, isAttachmentfile?: boolean): this;
+    setThumbnail(url?: string, isAttachmentfile?: boolean): this;
     getJson(): string;
     build(): Embed;
 }