+++ lib/services/MessageBuilder.d.ts
@@ -1,8 +1,9 @@
-import { DiscordMessage, Embed, Parse } from '../types';
+import { DiscordMessage, Embed, Parse, WebhookSettings, DiscordFile } from '../types';
 export default class MessageBuilder {
     private message;
+    overrideWebhook(settings?: WebhookSettings): this;
     setContent(content?: string): this;
-    addAttachment(attachment?: string): this;
+    addAttachment(attachment?: DiscordFile): this;
     addEmbed(embed?: Embed): this;
     setTTS(content?: string): this;
     setAllowedMentionsParse(parse?: Parse[]): this;